populate current user data in edit profile
