+++ src/components/Welcome.js
@@ -2,17 +2,17 @@ import WelcomeForm from './WelcomeForm';
 // import WelcomeSubmitBtn from './WelcomeSubmitBtn';
 import WelcomeSkipBtn from './WelcomeSkipBtn';
 import React, { Component } from 'react';
-
+import Container from 'react-bootstrap/Container';
 
 
 class Welcome extends Component {
   render() {
     return (
-      <div>
+      <Container className="pt-5">
         <h1>Welcome</h1>
         <WelcomeForm createUser={this.props.createUser} />
         <WelcomeSkipBtn />
-      </div>
+      </Container>
     );
   }
 }
